feat: support diy TableName
